Implement game win logic
Previously the game allowed adding to the board however it had
no way of tracking wins. I introduced some logic functions in
the Game function in order to check the column, row, and
diagonals of the previously placed mark. This allowed
to check a win without checking the entire board. Currently the
program only checks for wins, logic still needs to be applied to
check for a draw and to prevent players from adding to the board
once the game has already been won.